refactor: move time range logic to separate module
The `rewrite_expression` module was getting large, so made sense to
move time range logic to its own module.
